Refactor: 문구 및 변수명 수정
